requirements.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
gitdb==4.0.12
GitPython==3.1.45
idna==3.10
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
narwhals==2.5.0
numpy==2.3.3
packaging==24.2
pandas==2.3.2
pillow==11.3.0
protobuf==5.29.5
pyarrow==21.0.0
pydeck==0.9.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
requests==2.32.5
rich==13.9.4
rpds-py==0.27.1
six==1.17.0
smmap==5.0.2
streamlit==1.41.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0

# LangChain and AI packages
python-dotenv==1.1.1
langchain==0.3.27
langchain-openai==0.3.33
openai==1.109.1
langchain-core==0.3.76
pydantic==2.11.9
tiktoken==0.11.0
